% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/count_shm_motifs.R
\name{count_shm_motifs}
\alias{count_shm_motifs}
\title{Count the Number of Somatic Hypermutation (SHM) Motifs}
\usage{
count_shm_motifs(gr, bs.genome)
}
\arguments{
\item{gr}{GRanges object. There needs to be a id column in the metadata.
So that the number of motifs can be associated with the original
GenomicRange.}

\item{BS.genome}{This a Biostring-based genome object. (BSgenome from
Bioconductor). For instance, library("BSgenome.Hsapiens.UCSC.hg19")
can be used.}
}
\value{
A n x 3 data.frame indicating the original GRanges id, the number
}
\description{
Counts the number of the SHM motifs within the GenomicRange (gr) object.
}

