% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_alphabet_freq.R
\name{get_alphabet_freq}
\alias{get_alphabet_freq}
\title{Return the Alphabet Distribution Over a GRanges Object}
\usage{
get_alphabet_freq(gr, bs.genome)
}
\arguments{
\item{gr}{GRanges object. There needs to be a id column in the metadata.
So that the number of motifs can be associated with the original
GenomicRange.}

\item{BS.genome}{This a Biostring-based genome object. (BSgenome from
Bioconductor). For instance, library("BSgenome.Hsapiens.UCSC.hg19")
can be used.}
}
\value{
A matrix containing with letters as columns and each GRange as
        a row. Each value indicates the prevalence of that letter in the
        GRange
}
\description{
Calculates the number of different bases there are in a particlar GRanges
object.
}

