% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_shm_motifs.R
\name{get_shm_motifs}
\alias{get_shm_motifs}
\title{Get Somatic Hypermutation (SHM) Motif Locations}
\usage{
get_shm_motifs(gr, bs.genome)
}
\arguments{
\item{gr}{GenomicRanges object. There needs to be a id column in the metadata.
So that the motifs can be  associated with the original GenomicRange.}

\item{BS.genome}{This a Biostring-based genome object. (BSgenome from
Bioconductor). For instance, library("BSgenome.Hsapiens.UCSC.hg19")
can be used.}
}
\value{
An IRanges object with the relatives coordinates of the motifs in
        each GenomicRange.
}
\description{
Reports the location of the SHM motifs within the GenomicRange (gr) object.
}

