% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data-GRCh37.72.gr.R
\docType{data}
\name{gr}
\alias{gr}
\title{GenomicRanges of Chromosome 1 in GRCh37.72}
\format{\preformatted{Formal class 'GRanges' [package "GenomicRanges"] with 6 slots
  ..@ seqnames       :Formal class 'Rle' [package "S4Vectors"] with 4 slots
  .. .. ..@ values         : Factor w/ 1 level "chr1": 1
  .. .. ..@ lengths        : int 2071
  .. .. ..@ elementMetadata: NULL
  .. .. ..@ metadata       : list()
  ..@ ranges         :Formal class 'IRanges' [package "IRanges"] with 6 slots
  .. .. ..@ start          : int [1:2071] 39546988 19534771 9146538 2321147 1258072 915498 53049 69091 137380 367640 ...
  .. .. ..@ width          : int [1:2071] 2001 2001 2001 2001 2001 2001 2001 2001 2001 2001 ...
  .. .. ..@ NAMES          : NULL
  .. .. ..@ elementType    : chr "integer"
  .. .. ..@ elementMetadata: NULL
  .. .. ..@ metadata       : list()
  ..@ strand         :Formal class 'Rle' [package "S4Vectors"] with 4 slots
  .. .. ..@ values         : Factor w/ 3 levels "+","-","*": 1 2 1 2 1 2 1 2 1 2 ...
  .. .. ..@ lengths        : int [1:1007] 1 5 2 1 1 1 1 1 1 1 ...
  .. .. ..@ elementMetadata: NULL
  .. .. ..@ metadata       : list()
  ..@ elementMetadata:Formal class 'DataFrame' [package "S4Vectors"] with 6 slots
  .. .. ..@ rownames       : NULL
  .. .. ..@ nrows          : int 2071
  .. .. ..@ listData       :List of 1
  .. .. .. ..$ id: chr [1:2071] "MACF1" "UBR4" "SLC2A5" "MORN1" ...
  .. .. ..@ elementType    : chr "ANY"
  .. .. ..@ elementMetadata: NULL
  .. .. ..@ metadata       : list()
  ..@ seqinfo        :Formal class 'Seqinfo' [package "GenomeInfoDb"] with 4 slots
  .. .. ..@ seqnames   : chr "chr1"
  .. .. ..@ seqlengths : int NA
  .. .. ..@ is_circular: logi NA
  .. .. ..@ genome     : chr NA
  ..@ metadata       : list()
}}
\usage{
gr
}
\description{
A GenomicRanges object containing 2000 bp windows downstream of
the transcription start sites of all genes on chromsome 1
}
\keyword{datasets}

