% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shmutils.R
\docType{package}
\name{shmutils}
\alias{shmutils}
\alias{shmutils-package}
\title{shmutils: An R Package for Analyzing Somatic Hypermutation Patterns}
\description{
shmutils: An R Package for Analyzing Somatic Hypermutation Patterns
}

